backend-data-chart/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0